Move logic to help modal
